+++ openai-client/src/commonMain/kotlin/com.aallam.openai.client/internal/api/AudioApi.kt
@@ -9,7 +9,9 @@ import com.aallam.openai.client.Audio
 import com.aallam.openai.client.internal.extension.appendFileSource
 import com.aallam.openai.client.internal.http.HttpRequester
 import com.aallam.openai.client.internal.http.perform
+import com.aallam.openai.client.internal.http.performHeaders
 import io.ktor.client.request.forms.*
+import io.ktor.http.*
 
 /**
  * Implementation of [Audio].
@@ -23,8 +25,22 @@ internal class AudioApi(val requester: HttpRequester) : Audio {
         }
     }
 
+    @BetaOpenAI
+    override suspend fun transcriptionHeaders(request: TranscriptionRequest): Pair<Transcription, Headers> {
+        return when (request.responseFormat) {
+            "json", "verbose_json", null -> transcriptionAsJsonHeaders(request)
+            else -> transcriptionAsStringHeaders(request)
+        }
+    }
+
     private suspend fun transcriptionAsJson(request: TranscriptionRequest): Transcription {
         return requester.perform {
+            it.submitFormWithBinaryData(url = ApiPath.Transcription, formData = formDataOf(request))
+        }
+    }
+
+    private suspend fun transcriptionAsJsonHeaders(request: TranscriptionRequest): Pair<Transcription, Headers> {
+        return requester.performHeaders {
             it.submitFormWithBinaryData(url = ApiPath.Transcription, formData = formDataOf(request))
         }
     }
@@ -34,6 +50,13 @@ internal class AudioApi(val requester: HttpRequester) : Audio {
             it.submitFormWithBinaryData(url = ApiPath.Transcription, formData = formDataOf(request))
         }
         return Transcription(text)
+    }
+
+    private suspend fun transcriptionAsStringHeaders(request: TranscriptionRequest): Pair<Transcription, Headers> {
+        val text: Pair<String, Headers> = requester.performHeaders {
+            it.submitFormWithBinaryData(url = ApiPath.Transcription, formData = formDataOf(request))
+        }
+        return Transcription(text.first) to text.second
     }
 
     /**
@@ -57,6 +80,14 @@ internal class AudioApi(val requester: HttpRequester) : Audio {
         }
     }
 
+    @BetaOpenAI
+    override suspend fun translationHeaders(request: TranslationRequest): Pair<Translation, Headers> {
+        return when (request.responseFormat) {
+            "json", "verbose_json", null -> translationAsJsonHeaders(request)
+            else -> translationAsStringHeaders(request)
+        }
+    }
+
     private suspend fun translationAsJson(request: TranslationRequest): Translation {
         return requester.perform {
             it.submitFormWithBinaryData(url = ApiPath.Translation, formData = formDataOf(request))
@@ -68,6 +99,19 @@ internal class AudioApi(val requester: HttpRequester) : Audio {
             it.submitFormWithBinaryData(url = ApiPath.Translation, formData = formDataOf(request))
         }
         return Translation(text)
+    }
+
+    private suspend fun translationAsJsonHeaders(request: TranslationRequest): Pair<Translation, Headers> {
+        return requester.performHeaders {
+            it.submitFormWithBinaryData(url = ApiPath.Translation, formData = formDataOf(request))
+        }
+    }
+
+    private suspend fun translationAsStringHeaders(request: TranslationRequest): Pair<Translation, Headers> {
+        val text = requester.performHeaders<String> {
+            it.submitFormWithBinaryData(url = ApiPath.Translation, formData = formDataOf(request))
+        }
+        return Translation(text.first) to text.second
     }
 
     /**