+++ sample/jvm/src/main/kotlin/com/aallam/openai/sample/jvm/Chat.kt
@@ -36,3 +36,27 @@ suspend fun CoroutineScope.chat(openAI: OpenAI) {
         .launchIn(this)
         .join()
 }
+
+@OptIn(BetaOpenAI::class)
+suspend fun chatWithHeaders(openAI: OpenAI) {
+    println("\n> Create chat completions with headers...")
+    val chatCompletionRequest = ChatCompletionRequest(
+        model = ModelId("gpt-3.5-turbo"),
+        messages = listOf(
+            ChatMessage(
+                role = ChatRole.System,
+                content = "You are a helpful assistant that translates English to French."
+            ),
+            ChatMessage(
+                role = ChatRole.User,
+                content = "Translate the following English text to French: “OpenAI is awesome!”"
+            )
+        )
+    )
+    val response = openAI.chatCompletionHeaders(chatCompletionRequest)
+
+    val text = response.first.choices.firstOrNull()?.message?.content ?: ""
+    val headers = response.second
+
+    println("text: $text. headers: $headers")
+}